Exract SettingStoreBase
